Add prefix `input-` to id of all WTForms fields
The templates in digitalmarketplace-frontend-toolkit HTML input element
ids with `input-`. We need to know the id for form fields to be able to
link to them in error summaries.

This commit adds logic to our custom WTForms classes to prefix the id
with `input-`. This means that the WTForms object and the HTML template
should (fingers crossed) have the same id for the input in all cases,
whereas before you could have situations where we had to tell WTForms
the correct id.

To avoid this being a breaking change, we only add the input prefix if
the id is **not** specified. This should avoid situations where the id
would start with `input-input-`...