Add ability to override href for govuk_error object
We want to be able to use the dmutils.forms.errors helpers everywhere we
are using forms in our frontend apps. Sometimes our forms can be a
little... weird. Having the ability to override the field input href is
useful, especially for things like fields inside selection buttons.

This would be a breaking change due to the change in the href of
govuk_errors_from_wtforms, but for the previous commit. The effect of
these two commits together is to move the logic prefixing `input-` out
of the function and into the object.