Add suffix `-1` to id for selection button fields
In digitalmarketplace-frontend-toolkit selection buttons (radios and
checkboxes) have numbers after them to give each option a separate ID.

This means to be able to link to the selection buttons from the error
summary we need to also add a suffix to the ID in the WTForms objects.

This change means that a link in an error summary will link to the first
input element, mirroring the behaviour in govuk-frontend[1].

Note that in govuk-frontend v3 and above the first input element in a
radios component does not have a number suffix, so this code will not
work with govuk-frontend v3 and above[2] (not that it does anyway).

[1]: https://github.com/alphagov/govuk-frontend/blob/v3.7.0/src/govuk/components/radios/template.njk#L55
[2]: https://github.com/alphagov/govuk-frontend/commit/e25b0e100298ca3ed758bfda3f312448d84ca229